requirements.txt
addict==2.4.0
albumentations==1.3.0
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
defusedxml==0.7.1
easydict==1.10
faiss-cpu==1.7.3
fastapi==0.95.0
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==23.1.21
fonttools==4.38.0
h11==0.14.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
imageio==2.25.0
importlib-metadata==6.0.0
insightface==0.7.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jstyleson==0.0.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.3
mpmath==1.2.1
networkx==2.8.8
numpy==1.23.4
onnx==1.13.0
onnxruntime==1.13.1
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
openvino==2022.3.0
openvino-dev==2022.3.0
openvino-telemetry==2022.3.0
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
prettytable==3.6.0
protobuf==3.20.3
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2023.2.3
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.21.1
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.2
zipp==3.12.1
